PG: query_sub (generic trait bounds confusing)
