+++ Cargo.toml
@@ -45,3 +45,4 @@ chrono = "0.4.23"
 
 [dev-dependencies]
 anyhow = "1"
+nostr-sdk = "0.9.0"
